chore: refactor api code
