+++ api/src/app.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import { AppController } from './app.controller';
-import { NftController } from './nft.controller';
-import { AppService } from './app.service';
+import { AppController, NftController } from './controllers';
+import { AppService, NftService } from './services';
 
 @Module({
   // from root of app, not this file as base
   imports: [ConfigModule.forRoot({ envFilePath: '../.env' })],
   controllers: [AppController, NftController],
-  providers: [AppService],
+  providers: [AppService, NftService],
 })
 export class AppModule {}
