+++ api/src/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from './app.controller';
+export * from './nft.controller';
