+++ api/src/services/index.ts
@@ -1,0 +1,2 @@
+export * from './app.service';
+export * from './nft.service';
